user-actor-my-pets - move DTOs into lib/models
